requirements.txt
python-terrier>=0.11.0
pyterrier-adaptive>=0.1.0
pyterrier-caching>=0.1.1
pyterrier-alpha>=0.9.1
pyterrier-pisa>=0.2.0
